tiling重构

+++ mc2/allto_allv_grouped_mat_mul/op_kernel/mc2_templates/common/a2av_common_tiling.h
@@ -65,8 +65,8 @@ struct TaskTilingInfo {
     uint32_t totalLoopCount;     // 总循环次数
     
     // 通信参数（对应sendCounts和recvCounts）
-    int16_t sendCnt[MAX_EXPERT_SIZE];  // 每个expert的发送计数
-    int16_t recvCnt[MAX_EXPERT_SIZE];  // 每个expert的接收计数
+    int64_t sendCnt[MAX_EXPERT_SIZE];  // 每个expert的发送计数
+    int64_t recvCnt[MAX_EXPERT_SIZE];  // 每个expert的接收计数
 };
 
 }
